pyproject.toml
[tool.poetry]
name = "bacorex-sarl"
version = "0.1.0"
description = "Application web de gestion de projet de BACOREX-SARL"
authors = ["HAMA Barhamou <hamabarhamou@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.1.3"
Pillow = "9.3.0"
django-bootstrap-v5 = "1.0.11"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"
django-ckeditor = "^6.5.1"
djangorestframework = "^3.14.0"
pandas = "^2.1.0"
openpyxl = "^3.1.2"
black = "^23.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
